Added saving of temp variables.

+++ TEDExtract.py
@@ -14,6 +14,7 @@ import http
 from time import sleep
 import pandas as pd
 from bs4 import BeautifulSoup
+from six.moves import cPickle
 
 class TEDExtract(object):
   '''
@@ -36,11 +37,18 @@ class TEDExtract(object):
     Parse all pages and save their transcripts in own csv files.
     '''
 
-    # Collect all talks from all pages
-    for p in range(1, self.args.max_pages + 1):
-      path = 'https://www.ted.com/talks?page={}'.format(p)
+    if os.path.isfile(os.path.join(self.args.output, 'talk_list.pkl')):
+      with open(os.path.join(self.args.output, 'talk_list.pkl'), 'rb') as fin:
+        self.all_talks = cPickle.load(fin)
+    else:
+      # Collect all talks from all pages
+      for p in range(1, self.args.max_pages + 1):
+        path = 'https://www.ted.com/talks?page={}'.format(p)
 
-      self._fetch_talk_list(path)
+        self._fetch_talk_list(path)
+
+      with open(os.path.join(self.args.output, 'talk_list.pkl'), 'wb') as fout:
+        cPickle.dump(self.all_talks, fout)
 
     # DEBUG
     
@@ -71,43 +79,46 @@ class TEDExtract(object):
     # Extract the talk name
     talkname = talk[7:]
 
-    # The data frame object for saving all languages
-    req = urllib.request.urlopen('https://ted.com' + talk + '/transcript').read()
-    #conn.request('GET', talk + '/transcript', headers=hdr)
-    #req = conn.getresponse().read()
-    print("Search URL: {}".format(talk + '/transcript'))
-    print("Result: {}".format(req))
-    soup = BeautifulSoup(req)
+    if os.path.isfile(os.path.join(self.args.output, talkname + '.csv')):
+      print("Already there, skip {}".format(talk))
+    else:
+      # The data frame object for saving all languages
+      req = urllib.request.urlopen('https://ted.com' + talk + '/transcript').read()
+      #conn.request('GET', talk + '/transcript', headers=hdr)
+      #req = conn.getresponse().read()
+      print("Search URL: {}".format(talk + '/transcript'))
+      print("Result: {}".format(req))
+      soup = BeautifulSoup(req)
 
-    df = pd.DataFrame()
+      df = pd.DataFrame()
 
-    for i in soup.findAll('link'):
-      if i.get('href') != None and i.attrs['href'].find('?language=') != -1:
-        lang = i.attrs['hreflang']
-        path = i.attrs['href']
-        print("Path: {}".format(path))
-        #conn.request('GET', path, headers=hdr)
-        #r1 = conn.getresponse().read()
-        r1 = urllib.request.urlopen(path).read()
-        soup1 = BeautifulSoup(r1)
-        text_talk = []
-        for i in soup1.findAll('p',class_=''):
-          temo = i.text.replace('\n',' ')
-          temo = temo.replace('\t','')
-          temo = temo.replace('\r','')
-          temo = temo.replace('  ',' ')
+      for i in soup.findAll('link'):
+        if i.get('href') != None and i.attrs['href'].find('?language=') != -1:
+          lang = i.attrs['hreflang']
+          path = i.attrs['href']
+          print("Path: {}".format(path))
+          #conn.request('GET', path, headers=hdr)
+          #r1 = conn.getresponse().read()
+          r1 = urllib.request.urlopen(path).read()
+          soup1 = BeautifulSoup(r1)
+          text_talk = []
+          for i in soup1.findAll('p',class_=''):
+            temo = i.text.replace('\n',' ')
+            temo = temo.replace('\t','')
+            temo = temo.replace('\r','')
+            temo = temo.replace('  ',' ')
 
-          if temo[0] == ' ':
-            temo = temo[1:]
-          
-          text_talk.append(temo)
-        df1 = pd.DataFrame()
-        df1[lang] = text_talk
-        df = pd.concat([df,df1],axis=1)
+            if temo[0] == ' ':
+              temo = temo[1:]
+            
+            text_talk.append(temo)
+          df1 = pd.DataFrame()
+          df1[lang] = text_talk
+          df = pd.concat([df,df1],axis=1)
 
-        # This break is necessary, otherwise we'll receive a 429. THere would be
-        # a potential wordaround as noted by the commented out code snippets, but
-        # unfortunately this is not working
-        sleep(self.args.delay)
+          # This break is necessary, otherwise we'll receive a 429. THere would be
+          # a potential wordaround as noted by the commented out code snippets, but
+          # unfortunately this is not working
+          sleep(self.args.delay)
 
-    df.to_csv(os.path.join(self.args.output, talkname + '.csv'), sep='\t', encoding='utf-8')
+      df.to_csv(os.path.join(self.args.output, talkname + '.csv'), sep='\t', encoding='utf-8')
